Add WellBoreBaesComponent to SchematicLayer
